refactoring optimizers, updates to optimizer notebook
