+++ dspy/teleprompt/__init__.py
@@ -1,10 +1,11 @@
+from .teleprompt import *
 from .bootstrap import *
-from .ensemble import *
-from .finetune import *
-from .knn_fewshot import *
+from .vanilla import *
 from .random_search import *
+from .finetune import *
+from .teleprompt_optuna import *
+from .knn_fewshot import *
 from .signature_opt import SignatureOptimizer
 from .signature_opt_bayesian import BayesianSignatureOptimizer
-from .teleprompt import *
-from .teleprompt_optuna import *
-from .vanilla import *
+from .mipro_optimizer import MIPRO
+from .copro_optimizer import COPRO
